refactor: create card folder component
